reorganized & improved builder selection logic
