+++ builders/natdexoffensebuilder.js
@@ -1,6 +1,6 @@
-const leads = require("./mon-sets/leads.json");
-const sets = require("./mon-sets/sets.json");
-const config = require("./config.json");
+const leads = require("../mon-sets/leads.json");
+const sets = require("../mon-sets/sets.json");
+const config = require("../config.json");
 
 
 module.exports = buildTeam();
